Bring PX4-SITL_gazebo up-to-date with the upstream repo used by the official PX4-Autopilot:stable. (#2)
* fix imu noise model.

* Improve realtime factor for standard_vtol

* gimbal_controller: fix attitude status

This fixes two things that were implemented incorrectly:
- The flags were not set correctly.
- The quaternion was sent in the absolute frame even without yaw lock.

* Update cloudship.sdf.jinja

Replace GPS joint with statement including GPS joint defined in other GPS model folder. This is adapted from the method that other Gazebo SITL models include the GPS such as the plane. The base_link term is replaced by hull to attach the GPS to the hull.

* Revert "models: Add model for standard_vtol_ctrlalloc"

This reverts commit a3eef9d3a17078896f40abb0deaabfb0faad8c65.

* Fix HITL by sending heartbeats

PX4 no longer starts the mavlink instance on USB by default. This means
that the simulator needs to send heartbeats to initialize it.

Sending the heartbeats at 1 Hz however is not enough as we might fill up
all the buffers in less than a second which means PX4 never sees a
heartbeat, and won't initialize. Therefore, we send heartbeats in every
iteration initially until we have received a heartbeat back.

* resolved differences to match PX4

* reverted files in .github to match PX4

Co-authored-by: Shaochang Tan <478710209@qq.com>
Co-authored-by: Jaeyoung-Lim <jalim@ethz.ch>
Co-authored-by: Julian Oes <julian@oes.ch>
Co-authored-by: JamesAnawati <99291339+JamesAnawati@users.noreply.github.com>